gateway: record-live keeps full bodies (not the preview cap)
Record-live persists the live decode as the authoritative record, so it must keep
full bodies rather than the live preview (256 KiB). Two caps were in the way:

- The decoders cap decode.Flow bodies at maxLiveBody. Make it a process-wide var
  and add SetUnlimitedLiveBodies, called from server.Register when record-live is
  on — so the Go live decoders keep full bodies.
- The hub only stored the *proto* flow, whose bodies are dropped above the inline
  limit (and the proto stays a bounded preview for live streaming, on purpose).
  Retain the full decode.Flow per id in record-live mode (onFlow) and persist from
  those; the streamed proto is unchanged. WS message payloads are already inlined
  in full once the decoder cap is lifted.

recordlive_test.go now asserts a 700 KiB response body round-trips in full;
livebody_test.go covers the cap toggle.

+++ gateway/internal/decode/livehttp2.go
@@ -205,7 +205,7 @@ func maybeGunzip(raw []byte) []byte {
 		return raw // truncated/partial (capped) gzip — leave the raw bytes
 	}
 	defer zr.Close()
-	out, err := io.ReadAll(io.LimitReader(zr, maxLiveBody))
+	out, err := io.ReadAll(io.LimitReader(zr, int64(maxLiveBody)))
 	if err != nil && len(out) == 0 {
 		return raw
 	}
